@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c) 2020-2021 Mediatheken DLNA Bridge Authors.
+ * Copyright (c) 2020-2023 Mediatheken DLNA Bridge Authors.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -26,8 +26,8 @@ package de.corelogics.mediaview.repository.clip;
 
 import de.corelogics.mediaview.client.mediathekview.ClipEntry;
 import de.corelogics.mediaview.config.MainConfiguration;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
+import lombok.extern.log4j.Log4j2;
+import lombok.val;
 import org.apache.lucene.analysis.standard.StandardAnalyzer;
 import org.apache.lucene.document.*;
 import org.apache.lucene.facet.FacetsCollector;
@@ -58,6 +58,7 @@ import java.util.stream.Stream;
 import static java.util.Optional.ofNullable;
 import static java.util.stream.Collectors.toMap;
 
+@Log4j2
 public class ClipRepository {
     private enum ClipField {
         ID(true, false),
@@ -133,7 +134,6 @@ public class ClipRepository {
         T search(IndexSearcher searcher) throws IOException;
     }
 
-    private static final Logger LOGGER = LogManager.getLogger(ClipRepository.class);
     private static final String DOCID_LAST_UPDATED = "last-update-stat";
     private static final FieldType TYPE_NO_TOKENIZE = new FieldType();
 
@@ -153,13 +153,9 @@ public class ClipRepository {
         this.mainConfiguration = mainConfiguration;
         openConnection(calcIndexPath(), calcCacheSize());
         initialize();
-        LOGGER.info("Successfully opened database at {} with {} Bytes cache",
-                this::calcIndexPath,
-                this::calcCacheSize);
-    }
-
-    public void compact() {
-        // TODO not needed any more
+        log.info("Successfully opened database at {} with {} Bytes cache",
+            this::calcIndexPath,
+            this::calcCacheSize);
     }
 
     long calcCacheSize() {
@@ -171,15 +167,15 @@ public class ClipRepository {
     }
 
     void initialize() {
-        var indexPath = calcIndexPath();
+        val indexPath = calcIndexPath();
         if (!"<in-mem>".equals(indexPath)) {
-            LOGGER.debug("Looking for old clipdb entries in '{}'", indexPath);
+            log.debug("Looking for old clipdb entries in '{}'", indexPath);
             ofNullable(new File(indexPath).getParentFile().listFiles((dir, name) -> name.startsWith("clipdb.")))
-                    .stream()
-                    .flatMap(Stream::of)
-                    .peek(file -> LOGGER.info("Removing old clip database file '{}'", file))
-                    .filter(f -> !f.delete())
-                    .forEach(file -> LOGGER.warn("Could not remove old clip database file '{}'", file.getAbsolutePath()));
+                .stream()
+                .flatMap(Stream::of)
+                .peek(file -> log.info("Removing old clip database file '{}'", file))
+                .filter(f -> !f.delete())
+                .forEach(file -> log.warn("Could not remove old clip database file '{}'", file.getAbsolutePath()));
         }
 
     }
@@ -194,8 +190,8 @@ public class ClipRepository {
 
             try {
                 // need at least one entry for a reader to work on...
-                var writer = new IndexWriter(this.index, new IndexWriterConfig(new StandardAnalyzer()));
-                var document = new Document();
+                val writer = new IndexWriter(this.index, new IndexWriterConfig(new StandardAnalyzer()));
+                val document = new Document();
                 document.add(new Field(ClipField.ID.term(), ClipField.ID.term("placeholder"), TYPE_NO_TOKENIZE));
                 writer.updateDocument(new Term(ClipField.ID.term(), ClipField.ID.term("placeholder")), document);
                 writer.close();
@@ -203,9 +199,9 @@ public class ClipRepository {
                 // index got corrupted (or it's an old version). Delete index and re-index later.
                 this.index.close();
                 Files.walk(new File(indexPath).toPath())
-                        .sorted(Comparator.reverseOrder())
-                        .map(Path::toFile)
-                        .forEach(File::delete);
+                    .sorted(Comparator.reverseOrder())
+                    .map(Path::toFile)
+                    .forEach(File::delete);
                 openConnection(indexPath, cacheSize);
                 return;
             }
@@ -217,13 +213,9 @@ public class ClipRepository {
         }
     }
 
-    public void shutdown() {
-        // not needed for this implementation
-    }
-
     private <T> T withSearcher(SearchFunction<T> function) {
         try {
-            var searcher = searcherManager.acquire();
+            val searcher = searcherManager.acquire();
             try {
                 return function.search(searcher);
             } finally {
@@ -235,11 +227,11 @@ public class ClipRepository {
     }
 
     public Optional<ZonedDateTime> findLastFullImport() {
-        LOGGER.debug("finding last full import");
+        log.debug("finding last full import");
         return withSearcher(searcher -> {
-            var result = searcher.search(new TermQuery(new Term(ClipField.ID.term(), ClipField.ID.term(DOCID_LAST_UPDATED))), 1);
+            val result = searcher.search(new TermQuery(new Term(ClipField.ID.term(), ClipField.ID.term(DOCID_LAST_UPDATED))), 1);
             if (result.scoreDocs.length > 0) {
-                var doc = searcher.getIndexReader().document(result.scoreDocs[0].doc);
+                val doc = searcher.getIndexReader().storedFields().document(result.scoreDocs[0].doc);
                 return Optional.of(ZonedDateTime.parse(doc.get(ClipField.IMPORTEDAT.value())));
             }
             return Optional.empty();
@@ -247,17 +239,17 @@ public class ClipRepository {
     }
 
     public synchronized void updateLastFullImport(ZonedDateTime dateTime) {
-        LOGGER.debug("Updating last full import time to {}", dateTime);
+        log.debug("Updating last full import time to {}", dateTime);
         try {
-            var analyzer = new StandardAnalyzer();
-            var indexWriterConfig = new IndexWriterConfig(analyzer);
-            try (var writer = new IndexWriter(this.index, indexWriterConfig)) {
-                var d = new Document();
+            val analyzer = new StandardAnalyzer();
+            val indexWriterConfig = new IndexWriterConfig(analyzer);
+            try (val writer = new IndexWriter(this.index, indexWriterConfig)) {
+                val d = new Document();
                 addToDocument(d, ClipField.ID, DOCID_LAST_UPDATED);
                 addDateToDocument(d, ClipField.IMPORTEDAT, dateTime);
                 writer.updateDocument(
-                        new Term(ClipField.ID.term(), ClipField.ID.term(DOCID_LAST_UPDATED)),
-                        applyFacets(d));
+                    new Term(ClipField.ID.term(), ClipField.ID.term(DOCID_LAST_UPDATED)),
+                    applyFacets(d));
             }
             searcherManager.maybeRefreshBlocking();
         } catch (final IOException e) {
@@ -266,16 +258,16 @@ public class ClipRepository {
     }
 
     public List<String> findAllChannels() {
-        LOGGER.debug("Finding all channels");
+        log.debug("Finding all channels");
         return withSearcher(searcher -> {
-            var query = new MatchAllDocsQuery();
-            var state = new DefaultSortedSetDocValuesReaderState(searcher.getIndexReader(), ClipField.CHANNELNAME.facet());
-            var fc = new FacetsCollector();
+            val query = new MatchAllDocsQuery();
+            val state = new DefaultSortedSetDocValuesReaderState(searcher.getIndexReader(), ClipField.CHANNELNAME.facet(), null);
+            val fc = new FacetsCollector();
             FacetsCollector.search(searcher, query, 10000, fc);
-            var facets = new SortedSetDocValuesFacetCounts(state, fc);
+            val facets = new SortedSetDocValuesFacetCounts(state, fc);
             return Stream.of(facets.getTopChildren(10000, ClipField.CHANNELNAME.facet()).labelValues)
-                    .map(l -> l.label)
-                    .collect(Collectors.toList());
+                .map(l -> l.label)
+                .collect(Collectors.toList());
         });
     }
 
@@ -283,16 +275,16 @@ public class ClipRepository {
      * @return name to number of clips
      */
     public Map<String, Integer> findAllContainedIns(String channelName) {
-        LOGGER.debug("Finding all containedIns for channel '{}'", channelName);
+        log.debug("Finding all containedIns for channel '{}'", channelName);
         return withSearcher(searcher -> {
-            var query = new TermQuery(new Term(ClipField.CHANNELNAME.termLower(), ClipField.CHANNELNAME.termLower(channelName)));
-            var state = new DefaultSortedSetDocValuesReaderState(searcher.getIndexReader(), ClipField.CONTAINEDIN.facet());
-            var fc = new FacetsCollector();
+            val query = new TermQuery(new Term(ClipField.CHANNELNAME.termLower(), ClipField.CHANNELNAME.termLower(channelName)));
+            val state = new DefaultSortedSetDocValuesReaderState(searcher.getIndexReader(), ClipField.CONTAINEDIN.facet(), null);
+            val fc = new FacetsCollector();
             FacetsCollector.search(searcher, query, 10000, fc);
-            var facets = new SortedSetDocValuesFacetCounts(state, fc);
+            val facets = new SortedSetDocValuesFacetCounts(state, fc);
             return Stream.of(facets.getTopChildren(10000, ClipField.CONTAINEDIN.facet()).labelValues)
-                    .map(l -> Map.entry(l.label, l.value.intValue()))
-                    .collect(toMap(Map.Entry::getKey, Map.Entry::getValue));
+                .map(l -> Map.entry(l.label, l.value.intValue()))
+                .collect(toMap(Map.Entry::getKey, Map.Entry::getValue));
 
         });
     }
@@ -301,94 +293,94 @@ public class ClipRepository {
      * @return name to number of clips
      */
     public Map<String, Integer> findAllContainedIns(String channelName, String startingWith) {
-        LOGGER.debug("Finding all containedIns for channel '{}' starting with '{}'", channelName, startingWith);
+        log.debug("Finding all containedIns for channel '{}' starting with '{}'", channelName, startingWith);
         return withSearcher(searcher -> {
-            var query = new BooleanQuery.Builder()
-                    .add(new TermQuery(new Term(ClipField.CHANNELNAME.termLower(), ClipField.CHANNELNAME.termLower(channelName))), BooleanClause.Occur.MUST)
-                    .add(new PrefixQuery(new Term(ClipField.CONTAINEDIN.termLower(), ClipField.CONTAINEDIN.termLower(startingWith))), BooleanClause.Occur.MUST)
-                    .build();
-            var state = new DefaultSortedSetDocValuesReaderState(searcher.getIndexReader(), ClipField.CONTAINEDIN.facet());
-            var fc = new FacetsCollector();
+            val query = new BooleanQuery.Builder()
+                .add(new TermQuery(new Term(ClipField.CHANNELNAME.termLower(), ClipField.CHANNELNAME.termLower(channelName))), BooleanClause.Occur.MUST)
+                .add(new PrefixQuery(new Term(ClipField.CONTAINEDIN.termLower(), ClipField.CONTAINEDIN.termLower(startingWith))), BooleanClause.Occur.MUST)
+                .build();
+            val state = new DefaultSortedSetDocValuesReaderState(searcher.getIndexReader(), ClipField.CONTAINEDIN.facet(), null);
+            val fc = new FacetsCollector();
             FacetsCollector.search(searcher, query, 10000, fc);
-            var facets = new SortedSetDocValuesFacetCounts(state, fc);
+            val facets = new SortedSetDocValuesFacetCounts(state, fc);
             return Stream.of(facets.getTopChildren(10000, ClipField.CONTAINEDIN.facet()).labelValues)
-                    .map(l -> Map.entry(l.label, l.value.intValue()))
-                    .collect(toMap(Map.Entry::getKey, Map.Entry::getValue));
+                .map(l -> Map.entry(l.label, l.value.intValue()))
+                .collect(toMap(Map.Entry::getKey, Map.Entry::getValue));
         });
     }
 
     private ClipEntry clipEntryFromDocument(Document result) {
         return new ClipEntry(
-                result.get(ClipField.CHANNELNAME.value()),
-                result.get(ClipField.CONTAINEDIN.value()),
-                ZonedDateTime.parse(result.get(ClipField.BROADCASTEDAT.value())),
-                result.get(ClipField.TITLE.value()),
-                result.get(ClipField.DURATION.value()),
-                result.getField(ClipField.SIZE.value()).numericValue().longValue(),
-                result.get(ClipField.URL.value()),
-                result.get(ClipField.URL_HD.value()));
+            result.get(ClipField.CHANNELNAME.value()),
+            result.get(ClipField.CONTAINEDIN.value()),
+            ZonedDateTime.parse(result.get(ClipField.BROADCASTEDAT.value())),
+            result.get(ClipField.TITLE.value()),
+            result.get(ClipField.DURATION.value()),
+            result.getField(ClipField.SIZE.value()).numericValue().longValue(),
+            result.get(ClipField.URL.value()),
+            result.get(ClipField.URL_HD.value()));
     }
 
     public List<ClipEntry> findAllClips(String channelId, String containedIn) {
-        LOGGER.debug("Finding all clips for channel '{}' and containedIn '{}'", channelId, containedIn);
+        log.debug("Finding all clips for channel '{}' and containedIn '{}'", channelId, containedIn);
         return withSearcher(searcher -> {
-            var result = searcher.search(
-                    new BooleanQuery.Builder()
-                            .add(new TermQuery(new Term(ClipField.CONTAINEDIN.termLower(), ClipField.CONTAINEDIN.termLower(containedIn))), BooleanClause.Occur.MUST)
-                            .add(new TermQuery(new Term(ClipField.CHANNELNAME.termLower(), ClipField.CHANNELNAME.termLower(channelId))), BooleanClause.Occur.MUST)
-                            .build(),
-                    1000,
-                    new Sort(
-                            new SortField(ClipField.BROADCASTEDAT.sorted(), SortField.Type.LONG, true),
-                            new SortField(ClipField.TITLE.sorted(), SortField.Type.STRING)));
-            var clipEntries = new ArrayList<ClipEntry>(result.scoreDocs.length);
-            for (var doc : result.scoreDocs) {
-                clipEntries.add(clipEntryFromDocument(searcher.doc(doc.doc)));
+            val result = searcher.search(
+                new BooleanQuery.Builder()
+                    .add(new TermQuery(new Term(ClipField.CONTAINEDIN.termLower(), ClipField.CONTAINEDIN.termLower(containedIn))), BooleanClause.Occur.MUST)
+                    .add(new TermQuery(new Term(ClipField.CHANNELNAME.termLower(), ClipField.CHANNELNAME.termLower(channelId))), BooleanClause.Occur.MUST)
+                    .build(),
+                1000,
+                new Sort(
+                    new SortField(ClipField.BROADCASTEDAT.sorted(), SortField.Type.LONG, true),
+                    new SortField(ClipField.TITLE.sorted(), SortField.Type.STRING)));
+            val clipEntries = new ArrayList<ClipEntry>(result.scoreDocs.length);
+            for (val doc : result.scoreDocs) {
+                clipEntries.add(clipEntryFromDocument(searcher.storedFields().document(doc.doc)));
             }
             return clipEntries;
         });
     }
 
     public Optional<ClipEntry> findClipById(String id) {
-        LOGGER.debug("Finding clip for id '{}'", id);
+        log.debug("Finding clip for id '{}'", id);
         return withSearcher(searcher -> {
-            var result = searcher.search(new TermQuery(new Term(ClipField.ID.term(), ClipField.ID.term(id))), 1);
+            val result = searcher.search(new TermQuery(new Term(ClipField.ID.term(), ClipField.ID.term(id))), 1);
             if (result.scoreDocs.length > 0) {
-                return Optional.of(clipEntryFromDocument(searcher.doc(result.scoreDocs[0].doc)));
+                return Optional.of(clipEntryFromDocument(searcher.storedFields().document(result.scoreDocs[0].doc)));
             }
             return Optional.empty();
         });
     }
 
     public List<ClipEntry> findAllClipsForChannelBetween(String channelName, ZonedDateTime startDate, ZonedDateTime endDate) {
-        LOGGER.debug("Finding clips of channel '{}' between '{}' and '{}'", channelName, startDate, endDate);
+        log.debug("Finding clips of channel '{}' between '{}' and '{}'", channelName, startDate, endDate);
         return withSearcher(searcher -> {
-            var result = searcher.search(
-                    new BooleanQuery.Builder()
-                            .add(new TermQuery(new Term(ClipField.CHANNELNAME.termLower(), ClipField.CHANNELNAME.termLower(channelName))), BooleanClause.Occur.MUST)
-                            .add(NumericDocValuesField.newSlowRangeQuery(ClipField.BROADCASTEDAT.sorted(), startDate.toEpochSecond(), endDate.toEpochSecond()), BooleanClause.Occur.MUST)
-                            .build(),
-                    1000,
-                    new Sort(
-                            new SortField(ClipField.BROADCASTEDAT.sorted(), SortField.Type.LONG),
-                            new SortField(ClipField.TITLE.sorted(), SortField.Type.STRING)));
-            var clipEntries = new ArrayList<ClipEntry>(result.scoreDocs.length);
-            for (var doc : result.scoreDocs) {
-                clipEntries.add(clipEntryFromDocument(searcher.doc(doc.doc)));
+            val result = searcher.search(
+                new BooleanQuery.Builder()
+                    .add(new TermQuery(new Term(ClipField.CHANNELNAME.termLower(), ClipField.CHANNELNAME.termLower(channelName))), BooleanClause.Occur.MUST)
+                    .add(NumericDocValuesField.newSlowRangeQuery(ClipField.BROADCASTEDAT.sorted(), startDate.toEpochSecond(), endDate.toEpochSecond()), BooleanClause.Occur.MUST)
+                    .build(),
+                1000,
+                new Sort(
+                    new SortField(ClipField.BROADCASTEDAT.sorted(), SortField.Type.LONG),
+                    new SortField(ClipField.TITLE.sorted(), SortField.Type.STRING)));
+            val clipEntries = new ArrayList<ClipEntry>(result.scoreDocs.length);
+            for (val doc : result.scoreDocs) {
+                clipEntries.add(clipEntryFromDocument(searcher.storedFields().document(doc.doc)));
             }
             return clipEntries;
         });
     }
 
     public synchronized void deleteClipsImportedBefore(ZonedDateTime startedAt) {
-        LOGGER.debug("Deleting all clips not imported at {}", startedAt);
+        log.debug("Deleting all clips not imported at {}", startedAt);
         try {
-            var analyzer = new StandardAnalyzer();
-            var indexWriterConfig = new IndexWriterConfig(analyzer);
-            try (var writer = new IndexWriter(this.index, indexWriterConfig)) {
+            val analyzer = new StandardAnalyzer();
+            val indexWriterConfig = new IndexWriterConfig(analyzer);
+            try (val writer = new IndexWriter(this.index, indexWriterConfig)) {
                 writer.deleteDocuments(
-                        NumericDocValuesField.newSlowRangeQuery(
-                                ClipField.IMPORTEDAT.sorted(), Long.MIN_VALUE, startedAt.toEpochSecond() - 1));
+                    NumericDocValuesField.newSlowRangeQuery(
+                        ClipField.IMPORTEDAT.sorted(), Long.MIN_VALUE, startedAt.toEpochSecond() - 1));
             }
             searcherManager.maybeRefreshBlocking();
         } catch (final IOException e) {
@@ -397,14 +389,14 @@ public class ClipRepository {
     }
 
     public synchronized void addClips(Iterable<ClipEntry> clipEntries, ZonedDateTime importedAt) {
-        LOGGER.debug("Adding ClipEntries");
+        log.debug("Adding ClipEntries");
         try {
-            var analyzer = new StandardAnalyzer();
-            var indexWriterConfig = new IndexWriterConfig(analyzer);
-            try (var writer = new IndexWriter(this.index, indexWriterConfig)) {
-                for (var e : clipEntries) {
-                    LOGGER.debug("Updating document with id '{}': '{}'", e.getId(), e.getTitle());
-                    var d = new Document();
+            val analyzer = new StandardAnalyzer();
+            val indexWriterConfig = new IndexWriterConfig(analyzer);
+            try (val writer = new IndexWriter(this.index, indexWriterConfig)) {
+                for (val e : clipEntries) {
+                    log.debug("Updating document with id '{}': '{}'", e.getId(), e.getTitle());
+                    val d = new Document();
                     addToDocument(d, ClipField.ID, e.getId());
                     addToDocument(d, ClipField.CHANNELNAME, e.getChannelName());
                     addToDocument(d, ClipField.CONTAINEDIN, e.getContainedIn());
@@ -416,8 +408,8 @@ public class ClipRepository {
                     addDateToDocument(d, ClipField.BROADCASTEDAT, e.getBroadcastedAt());
                     addDateToDocument(d, ClipField.IMPORTEDAT, importedAt);
                     writer.updateDocument(
-                            new Term(ClipField.ID.term(), ClipField.ID.term(e.getId())),
-                            applyFacets(d));
+                        new Term(ClipField.ID.term(), ClipField.ID.term(e.getId())),
+                        applyFacets(d));
                 }
             }
             searcherManager.maybeRefreshBlocking();
@@ -427,8 +419,8 @@ public class ClipRepository {
     }
 
     private Document applyFacets(Document d) throws IOException {
-        var facetsConfig = new FacetsConfig();
-        for (final var ixf : d) {
+        val facetsConfig = new FacetsConfig();
+        for (val ixf : d) {
             if (ixf.fieldType() == SortedSetDocValuesFacetField.TYPE) {
                 SortedSetDocValuesFacetField facetField = (SortedSetDocValuesFacetField) ixf;
                 facetsConfig.setIndexFieldName(facetField.dim, facetField.dim);
