@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c) 2020-2021 Mediatheken DLNA Bridge Authors.
+ * Copyright (c) 2020-2023 Mediatheken DLNA Bridge Authors.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -25,13 +25,19 @@
 package de.corelogics.mediaview.client.mediathekview;
 
 import de.corelogics.mediaview.util.HashingUtils;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.ToString;
 
 import java.time.ZonedDateTime;
-import java.util.Objects;
-import java.util.StringJoiner;
 
+@Getter
+@EqualsAndHashCode
+@ToString
 public class ClipEntry {
+    @EqualsAndHashCode.Exclude
     private final String id;
+
     private final String title;
     private final String containedIn;
     private final ZonedDateTime broadcastedAt;
@@ -57,79 +63,7 @@ public class ClipEntry {
         return HashingUtils.idHash(this.channelName, getBestUrl());
     }
 
-    public String getTitle() {
-        return title;
-    }
-
-    public String getContainedIn() {
-        return containedIn;
-    }
-
-    public ZonedDateTime getBroadcastedAt() {
-        return broadcastedAt;
-    }
-
-    public String getChannelName() {
-        return channelName;
-    }
-
-    public long getSize() {
-        return size;
-    }
-
-    public String getUrl() {
-        return url;
-    }
-
-    public String getUrlHd() {
-        return urlHd;
-    }
-
     public String getBestUrl() {
         return null != urlHd && !urlHd.isBlank() ? urlHd : url;
     }
-
-    public String getDuration() {
-        return duration;
-    }
-
-    public String getId() {
-        return this.id;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (!(o instanceof ClipEntry)) return false;
-        ClipEntry clipEntry = (ClipEntry) o;
-        return size == clipEntry.size &&
-                Objects.equals(title, clipEntry.title) &&
-                Objects.equals(containedIn, clipEntry.containedIn) &&
-                Objects.equals(broadcastedAt.toEpochSecond(), clipEntry.broadcastedAt.toEpochSecond()) &&
-                Objects.equals(channelName, clipEntry.channelName) &&
-                Objects.equals(url, clipEntry.url) &&
-                Objects.equals(urlHd, clipEntry.urlHd) &&
-                Objects.equals(duration, clipEntry.duration);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(title, containedIn, broadcastedAt.toEpochSecond(), channelName, size, url, urlHd, duration);
-    }
-
-    @Override
-    public String toString() {
-        return new StringJoiner(", ", ClipEntry.class.getSimpleName() + "[", "]")
-                .add("id='" + id + "'")
-                .add("title='" + title + "'")
-                .add("containedIn='" + containedIn + "'")
-                .add("broadcastedAt=" + broadcastedAt)
-                .add("channelName='" + channelName + "'")
-                .add("size=" + size)
-                .add("url='" + url + "'")
-                .add("urlHd='" + urlHd + "'")
-                .add("duration='" + duration + "'")
-                .toString();
-    }
-
 }
