@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c) 2020-2021 Mediatheken DLNA Bridge Authors.
+ * Copyright (c) 2020-2023 Mediatheken DLNA Bridge Authors.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -24,23 +24,22 @@
 
 package de.corelogics.mediaview.config;
 
+import lombok.Getter;
+
 import java.io.File;
 import java.io.IOException;
 
+@Getter
 public class ConfigurationModule {
     private final MainConfiguration mainConfiguration;
 
     public ConfigurationModule() throws IOException {
         this.mainConfiguration = new MainConfiguration(
-                new TypedConfigurationAccessor(
-                        new LayeredPropertySource(
-                                new PropertiesConfigurationSource(System.getProperties()),
-                                new MapConfigurationSource(System.getenv()),
-                                new PropertiesConfigurationSource(new File("./config/application.properties")),
-                                new PropertiesConfigurationSource(ConfigurationModule.class.getResource("/application.properties")))));
-    }
-
-    public MainConfiguration getMainConfiguration() {
-        return mainConfiguration;
+            new TypedConfigurationAccessor(
+                new LayeredPropertySource(
+                    new PropertiesConfigurationSource(System.getProperties()),
+                    new MapConfigurationSource(System.getenv()),
+                    new PropertiesConfigurationSource(new File("./config/application.properties")),
+                    new PropertiesConfigurationSource(ConfigurationModule.class.getResource("/application.properties")))));
     }
 }
