@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c) 2020-2021 Mediatheken DLNA Bridge Authors.
+ * Copyright (c) 2020-2023 Mediatheken DLNA Bridge Authors.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -24,35 +24,12 @@
 
 package de.corelogics.mediaview.config;
 
-import java.util.StringJoiner;
-
-public class FavouriteShow implements Favourite {
-    private final String channel;
-    private final String title;
-
-    public FavouriteShow(String channel, String title) {
-        this.channel = channel;
-        this.title = title;
-    }
-
+public record FavouriteShow(
+    String channel,
+    String title
+) implements Favourite {
     @Override
     public <T> T accept(FavouriteVisitor<T> visitor) {
         return visitor.visitShow(this);
-    }
-
-    public String getChannel() {
-        return channel;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    @Override
-    public String toString() {
-        return new StringJoiner(", ", FavouriteShow.class.getSimpleName() + "[", "]")
-                .add("channel='" + channel + "'")
-                .add("title='" + title + "'")
-                .toString();
     }
 }
