@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c) 2020-2021 Mediatheken DLNA Bridge Authors.
+ * Copyright (c) 2020-2023 Mediatheken DLNA Bridge Authors.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -24,50 +24,55 @@
 
 package de.corelogics.mediaview.service.dlna;
 
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-import org.fourthline.cling.support.contentdirectory.AbstractContentDirectoryService;
-import org.fourthline.cling.support.contentdirectory.ContentDirectoryErrorCode;
-import org.fourthline.cling.support.contentdirectory.ContentDirectoryException;
-import org.fourthline.cling.support.contentdirectory.DIDLParser;
-import org.fourthline.cling.support.model.BrowseFlag;
-import org.fourthline.cling.support.model.BrowseResult;
-import org.fourthline.cling.support.model.DIDLContent;
-import org.fourthline.cling.support.model.SortCriterion;
+import de.corelogics.mediaview.service.dlna.jupnp.LocalAddressHolder;
+import lombok.AllArgsConstructor;
+import lombok.extern.log4j.Log4j2;
+import lombok.val;
+import org.jupnp.support.contentdirectory.AbstractContentDirectoryService;
+import org.jupnp.support.contentdirectory.ContentDirectoryErrorCode;
+import org.jupnp.support.contentdirectory.ContentDirectoryException;
+import org.jupnp.support.contentdirectory.DIDLParser;
+import org.jupnp.support.model.BrowseFlag;
+import org.jupnp.support.model.BrowseResult;
+import org.jupnp.support.model.DIDLContent;
+import org.jupnp.support.model.SortCriterion;
 
 import java.util.Arrays;
 import java.util.Collection;
 
+@AllArgsConstructor
+@Log4j2
 class ContentDirectory extends AbstractContentDirectoryService {
-    private final Logger logger = LogManager.getLogger(ContentDirectory.class);
-
     private final Collection<DlnaRequestHandler> handlers;
 
-    public ContentDirectory(Collection<DlnaRequestHandler> handlers) {
-        this.handlers = handlers;
-    }
-
     @Override
-    public BrowseResult browse(String objectID, BrowseFlag browseFlag,
-                               String filter,
-                               long firstResult, long maxResults,
-                               SortCriterion[] orderby)
-            throws ContentDirectoryException {
-        logger.debug("Received browse request for oid={}, first={}, max={}", objectID, firstResult, maxResults);
+    public BrowseResult browse(
+        String objectID,
+        BrowseFlag browseFlag,
+        String filter,
+        long firstResult, long maxResults,
+        SortCriterion[] orderBy)
+        throws ContentDirectoryException {
+        log.debug("Received browse request for oid={}, first={}, max={}", objectID, firstResult, maxResults);
         try {
-            var request = new DlnaRequest(
-                    objectID, browseFlag, filter, firstResult, maxResults,
-                    Arrays.asList(orderby));
+            val request = new DlnaRequest(
+                objectID,
+                browseFlag,
+                filter,
+                firstResult,
+                maxResults,
+                Arrays.asList(orderBy),
+                LocalAddressHolder.getMemoizedLocalAddress());
             return handlers.stream()
-                    .filter(h -> h.canHandle(request))
-                    .findAny()
-                    .map(h -> h.respond(request))
-                    .orElseGet(this::emptyResult);
+                .filter(h -> h.canHandle(request))
+                .findAny()
+                .map(h -> h.respond(request))
+                .orElseGet(this::emptyResult);
         } catch (RuntimeException e) {
-            logger.warn("Error creating a browse response", e);
+            log.warn("Error creating a browse response", e);
             throw new ContentDirectoryException(
-                    ContentDirectoryErrorCode.CANNOT_PROCESS,
-                    e.toString());
+                ContentDirectoryErrorCode.CANNOT_PROCESS,
+                e.toString());
         }
     }
 
