@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c) 2020-2021 Mediatheken DLNA Bridge Authors.
+ * Copyright (c) 2020-2023 Mediatheken DLNA Bridge Authors.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -24,55 +24,19 @@
 
 package de.corelogics.mediaview.service.dlna;
 
-import org.fourthline.cling.support.model.BrowseFlag;
-import org.fourthline.cling.support.model.SortCriterion;
+import org.jupnp.support.model.BrowseFlag;
+import org.jupnp.support.model.SortCriterion;
 
+import java.net.InetAddress;
 import java.util.List;
+import java.util.Optional;
 
-public class DlnaRequest {
-    private final String objectId;
-
-    private final BrowseFlag browseFlag;
-
-    private final String filer;
-
-    private final long firstResult;
-
-    private final long maxResults;
-
-    private final List<SortCriterion> orderBy;
-
-    public DlnaRequest(String objectId, BrowseFlag browseFlag, String filer, long firstResult, long maxResults,
-                       List<SortCriterion> orderBy) {
-        this.objectId = objectId;
-        this.browseFlag = browseFlag;
-        this.filer = filer;
-        this.firstResult = firstResult;
-        this.maxResults = maxResults;
-        this.orderBy = orderBy;
-    }
-
-    public String getObjectId() {
-        return objectId;
-    }
-
-    public BrowseFlag getBrowseFlag() {
-        return browseFlag;
-    }
-
-    public String getFiler() {
-        return filer;
-    }
-
-    public long getFirstResult() {
-        return firstResult;
-    }
-
-    public long getMaxResults() {
-        return maxResults;
-    }
-
-    public List<SortCriterion> getOrderBy() {
-        return orderBy;
-    }
+public record DlnaRequest(
+    String objectId,
+    BrowseFlag browseFlag,
+    String filer,
+    long firstResult,
+    long maxResults,
+    List<SortCriterion> orderBy,
+    Optional<InetAddress> localAddress) {
 }
