@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c) 2020-2021 Mediatheken DLNA Bridge Authors.
+ * Copyright (c) 2020-2023 Mediatheken DLNA Bridge Authors.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -24,7 +24,7 @@
 
 package de.corelogics.mediaview.service.dlna;
 
-import org.fourthline.cling.support.model.BrowseResult;
+import org.jupnp.support.model.BrowseResult;
 
 public interface DlnaRequestHandler {
     boolean canHandle(DlnaRequest request);
