@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c) 2020-2021 Mediatheken DLNA Bridge Authors.
+ * Copyright (c) 2020-2023 Mediatheken DLNA Bridge Authors.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -26,10 +26,13 @@ package de.corelogics.mediaview.service.proxy;
 
 import de.corelogics.mediaview.client.mediathekview.ClipEntry;
 import de.corelogics.mediaview.service.ClipContentUrlGenerator;
+import org.jetbrains.annotations.Nullable;
+
+import java.net.InetAddress;
 
 public class DirectDownloadClipContentUrlGenerator implements ClipContentUrlGenerator {
     @Override
-    public String createLinkTo(ClipEntry e) {
+    public String createLinkTo(ClipEntry e, @Nullable InetAddress optionalLocalAddressQueried) {
         return e.getBestUrl();
     }
 }
