@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c) 2020-2021 Mediatheken DLNA Bridge Authors.
+ * Copyright (c) 2020-2023 Mediatheken DLNA Bridge Authors.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -28,59 +28,70 @@ import com.fasterxml.jackson.core.JsonFactory;
 import com.github.benmanes.caffeine.cache.Caffeine;
 import com.github.benmanes.caffeine.cache.LoadingCache;
 import com.github.benmanes.caffeine.cache.RemovalCause;
+import com.github.benmanes.caffeine.cache.Ticker;
 import de.corelogics.mediaview.config.MainConfiguration;
 import de.corelogics.mediaview.util.IdUtils;
+import lombok.extern.log4j.Log4j2;
+import lombok.val;
 import org.apache.commons.io.IOUtils;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
+import org.jetbrains.annotations.NotNull;
 
 import java.io.*;
 import java.util.Arrays;
 import java.util.Comparator;
 import java.util.Optional;
 import java.util.Set;
-import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
+import static java.lang.Thread.ofVirtual;
 import static java.util.Optional.ofNullable;
+import static java.util.concurrent.Executors.newSingleThreadScheduledExecutor;
 
+@Log4j2
 public class CacheDirectory {
-    private final Logger logger = LogManager.getLogger(CacheDirectory.class);
-    private final ScheduledExecutorService scheduler = Executors.newSingleThreadScheduledExecutor();
     private final JsonFactory factory = new JsonFactory();
-
     private final File cacheDirFile;
     private final long cacheSizeBytes;
+    private final ScheduledExecutorService scheduledExecutorService = newSingleThreadScheduledExecutor(ofVirtual().factory());
+    private final LoadingCache<String, RandomAccessFile> openFiles;
 
-    private LoadingCache<String, RandomAccessFile> openFiles = Caffeine.newBuilder()
+    CacheDirectory(int cacheSizeGb, File cacheDir, Ticker cacheTicker) {
+        this.openFiles = Caffeine.newBuilder()
             .maximumSize(40)
             .expireAfterAccess(120, TimeUnit.SECONDS)
             .removalListener(this::closeFile)
+            .executor(Runnable::run)
+            .ticker(cacheTicker)
             .build(this::openFile);
 
+        if (cacheSizeGb < 10) {
+            throw new IllegalStateException("Configuration: CACHE_SIZE_GB is " + cacheSizeGb + ", but at least 10 GB are required");
+        }
+        this.cacheSizeBytes = 1024L * 1024L * 1024L * cacheSizeGb;
+        this.cacheDirFile = cacheDir;
+        log.debug("Initializing cache download manager, with cache in directory [{}]", this.cacheDirFile::getAbsolutePath);
+        if (!cacheDirFile.exists() && !cacheDirFile.mkdirs()) {
+            throw new IllegalStateException("Could not create nonexistent cache directory at " + this.cacheDirFile.getAbsolutePath());
+        }
+        scheduledExecutorService.scheduleAtFixedRate(this::cleanUp, 10, 10, TimeUnit.SECONDS);
+        log.info("Successfully started cache download manager, with cache in directory [{}]", this.cacheDirFile::getAbsolutePath);
+    }
 
     public CacheDirectory(MainConfiguration mainConfiguration) {
-        if (mainConfiguration.cacheSizeGb() < 10) {
-            throw new IllegalStateException("Configuration: CACHE_SIZE_GB is " + mainConfiguration.cacheSizeGb() + ", but at least 10 GB are required");
-        }
-        this.cacheSizeBytes = 1024L * 1024L * 1024L * mainConfiguration.cacheSizeGb();
-        this.cacheDirFile = new File(mainConfiguration.cacheDir());
-        logger.debug("Initializing cache download manager, with cache in directory [{}]", this.cacheDirFile::getAbsolutePath);
-        if (!cacheDirFile.exists() && !cacheDirFile.mkdirs()) {
-            throw new IllegalStateException("Could not create nonexisting cache directory at " + this.cacheDirFile.getAbsolutePath());
-        }
-        this.scheduler.scheduleAtFixedRate(this.openFiles::cleanUp, 10, 10, TimeUnit.SECONDS);
-        logger.info("Successfully started cache download manager, with cache in directory [{}]", this.cacheDirFile::getAbsolutePath);
+        this(mainConfiguration.cacheSizeGb(), mainConfiguration.cacheDir(), Ticker.systemTicker());
     }
 
     private RandomAccessFile openFile(String filename) throws FileNotFoundException {
+        log.debug("Opening file " + filename);
         return new RandomAccessFile(new File(cacheDirFile, filename), "rw");
     }
 
     private void closeFile(String name, RandomAccessFile file, RemovalCause cause) {
-        IOUtils.closeQuietly(file, e -> logger.debug("Could not (quietly) close file.", e));
+        log.debug("Closing file " + name);
+        IOUtils.closeQuietly(file, e -> log.debug("Could not (quietly) close file.", e));
     }
 
     private String contentFilename(String clipId) {
@@ -94,111 +105,151 @@ public class CacheDirectory {
 
     private long directorySize() {
         return ofNullable(this.cacheDirFile.listFiles()).stream()
-                .flatMap(Arrays::stream)
-                .filter(File::isFile)
-                .mapToLong(File::length)
-                .sum();
+            .flatMap(Arrays::stream)
+            .filter(File::isFile)
+            .mapToLong(File::length)
+            .sum();
     }
 
-    public synchronized void writeContent(String clipId, long position, byte[] bytes) throws IOException {
-        var contentFile = this.openFiles.get(contentFilename(clipId));
+    public void writeContent(String clipId, long position, byte[] bytes) throws IOException {
+        val contentFilename = contentFilename(clipId);
+        val contentFile = this.openFiles.get(contentFilename);
         synchronized (contentFile) {
+            if (position + bytes.length > contentFile.length()) {
+                throw new EOFException(
+                    "%s: write %d-%d > %d".formatted(
+                        contentFilename, position, position + bytes.length, contentFile.length()));
+            }
             contentFile.seek(position);
             contentFile.write(bytes);
         }
     }
 
-    public synchronized int readContentByte(String clipId, long position) throws IOException {
-        var contentFilename = contentFilename(clipId);
-        var contentFile = new File(this.cacheDirFile, contentFilename);
-        if (!contentFile.exists()) {
-            throw new FileNotFoundException(contentFilename);
+    private Optional<RandomAccessFile> openIfPresent(String filename) {
+        val existingContentAccess = this.openFiles.getIfPresent(filename);
+        if (null != existingContentAccess) {
+            return Optional.of(existingContentAccess);
         }
-
-        var contentAccess = this.openFiles.get(contentFilename);
-        if (position >= contentAccess.length()) {
-            throw new EOFException(contentFilename + ": " + position + " >" + contentFile.length());
+        // opening the nonexistent file would create it
+        val contentFile = new File(this.cacheDirFile, filename);
+        if (contentFile.exists()) {
+            return Optional.of(this.openFiles.get(filename));
         }
-        contentAccess.seek(position);
-        return contentAccess.read();
+        return Optional.empty();
     }
 
-    public synchronized int readContentBytes(String clipId, long position, byte[] buffer, int off, int len) throws IOException {
-        var contentFilename = contentFilename(clipId);
-        var contentFile = new File(this.cacheDirFile, contentFilename);
-        if (!contentFile.exists()) {
-            throw new FileNotFoundException(contentFilename);
+    public int readContentByte(String clipId, long position) throws IOException {
+        val contentFilename = contentFilename(clipId);
+        val contentAccess = this.openIfPresent(contentFilename)
+            .orElseThrow(() -> new FileNotFoundException(contentFilename));
+        synchronized (contentAccess) {
+            if (position >= contentAccess.length()) {
+                throw new EOFException(
+                    "%s: %d >%d".formatted(
+                        contentFilename, position, contentAccess.length()));
+            }
+            contentAccess.seek(position);
+            return contentAccess.read();
         }
+    }
 
-        var contentAccess = this.openFiles.get(contentFilename);
-        if (position >= contentAccess.length()) {
-            return -1;
+    public int readContentBytes(String clipId, long position, byte[] buffer, int off, int len) throws IOException {
+        val contentFilename = contentFilename(clipId);
+        val contentAccess = this.openIfPresent(contentFilename)
+            .orElseThrow(() -> new FileNotFoundException(contentFilename));
+        synchronized (contentAccess) {
+            if (position >= contentAccess.length()) {
+                throw new EOFException(
+                    "%s: %d > %d".formatted(contentFilename, position, contentAccess.length()));
+            }
+            contentAccess.seek(position);
+            val bytesLeftInFile = contentAccess.length() - position;
+            return contentAccess.read(buffer, off, (int) Math.min(bytesLeftInFile, len));
         }
-        contentAccess.seek(position);
-        var bytesLeftInFile = contentAccess.length() - position;
-        return contentAccess.read(buffer, off, (int) Math.min(bytesLeftInFile, len));
     }
 
     public synchronized void growContentFile(String clipId, long newSize) throws IOException, CacheSizeExhaustedException {
-        var contentFilename = contentFilename(clipId);
-        var contentFile = this.openFiles.get(contentFilename);
-        var dirSize = directorySize();
-        var contentFileSizePrior = contentFile.length();
-        var dirSizeAfter = dirSize - contentFileSizePrior + newSize;
-        if (dirSizeAfter > this.cacheSizeBytes) {
-            throw new CacheSizeExhaustedException(
+        val contentFilename = contentFilename(clipId);
+        val contentAccess = this.openFiles.get(contentFilename);
+        synchronized (contentAccess) {
+            val dirSize = directorySize();
+            val contentFileSizePrior = contentAccess.length();
+            val dirSizeAfter = dirSize - contentFileSizePrior + newSize;
+            if (dirSizeAfter > this.cacheSizeBytes) {
+                throw new CacheSizeExhaustedException(
                     String.format(
-                            "growing [%s] to [%d] would exceed cache size limit of %d with new size of %d",
-                            contentFilename,
-                            newSize,
-                            this.cacheSizeBytes,
-                            dirSizeAfter));
+                        "growing [%s] to [%d] would exceed cache size limit of %d with new size of %d",
+                        contentFilename,
+                        newSize,
+                        this.cacheSizeBytes,
+                        dirSizeAfter));
+            }
+            contentAccess.setLength(newSize);
         }
-        contentFile.setLength(newSize);
     }
 
 
-    public synchronized void writeMetadata(String clipId, ClipMetadata metadata) throws IOException {
-        var metaFile = this.openFiles.get(metaFilename(clipId));
-        metaFile.seek(0);
-        try (var generator = factory.createGenerator(metaFile)) {
-            metadata.writeTo(generator);
+    void writeMetadata(String clipId, @NotNull ClipMetadata metadata) throws IOException {
+        val metaFile = this.openFiles.get(metaFilename(clipId));
+        synchronized (metaFile) {
+            metaFile.seek(0);
+            try (val generator = factory.createGenerator(metaFile)) {
+                metadata.writeTo(generator);
+            }
+            metaFile.setLength(metaFile.getFilePointer());
         }
-        metaFile.setLength(metaFile.getFilePointer());
     }
 
-    public synchronized Optional<ClipMetadata> loadMetadata(String clipId) throws IOException {
-        var metaFile = new File(this.cacheDirFile, metaFilename(clipId));
-        if (!metaFile.exists()) {
+    Optional<ClipMetadata> loadMetadata(String clipId) throws IOException {
+        val metaFileAccessOpt = openIfPresent(metaFilename(clipId));
+        if (metaFileAccessOpt.isPresent()) {
+            val metaFileAccess = metaFileAccessOpt.get();
+            metaFileAccess.seek(0);
+            return Optional.of(ClipMetadata.readFrom(factory.createParser(metaFileAccess)));
+        } else {
             return Optional.empty();
         }
-        var metaFileAccess = this.openFiles.get(metaFilename(clipId));
-        metaFileAccess.seek(0);
-        return Optional.of(ClipMetadata.readFrom(factory.createParser(metaFileAccess)));
     }
 
     public synchronized boolean tryCleanupCacheDir(Set<String> currentlyOpenClipIds) {
-        var currentlyOpenFilenames = currentlyOpenClipIds.stream().map(this::contentFilename).collect(Collectors.toSet());
-
-        logger.info("Cleaning up some space in cache directory");
+        val currentlyOpenFilenames = currentlyOpenClipIds.stream()
+            .flatMap(clipId ->Stream.of(this.contentFilename(clipId), this.metaFilename(clipId)))
+            .collect(Collectors.toSet());
+        log.info("Cleaning up some space in cache directory");
         return ofNullable(this.cacheDirFile
-                .listFiles((dir, name) -> name.endsWith(".mp4"))).stream()
-                .flatMap(Arrays::stream)
-                .filter(f -> !currentlyOpenFilenames.contains(f.getName()))
-                .sorted(Comparator.comparing(File::lastModified))
-                .map(contentFileToRemove -> {
-                    var metaFileToRemove = new File(this.cacheDirFile, contentFileToRemove.getName().replaceAll(".mp4", ".json"));
-                    logger.debug(
-                            "Trying to delete {}, and {} with size of {}",
-                            metaFileToRemove::getName,
-                            contentFileToRemove::getName,
-                            () -> metaFileToRemove.length() + contentFileToRemove.length());
-                    this.openFiles.invalidate(metaFileToRemove.getName());
-                    this.openFiles.invalidate(contentFileToRemove.getName());
-                    return metaFileToRemove.delete() && contentFileToRemove.delete();
-                })
-                .filter(Boolean.TRUE::equals)
-                .findFirst()
-                .isPresent();
+            .listFiles((dir, name) -> name.endsWith(".mp4") || name.endsWith(".json"))).stream()
+            .flatMap(Arrays::stream)
+            .filter(f -> !currentlyOpenFilenames.contains(f.getName()))
+            .sorted(Comparator.comparing(File::lastModified))
+            .map(foundFilename -> {
+                val contentFileToRemove = new File(this.cacheDirFile, foundFilename.getName().replaceAll(".json", ".mp4"));
+                val metaFileToRemove = new File(this.cacheDirFile, contentFileToRemove.getName().replaceAll(".mp4", ".json"));
+                log.debug(
+                    "Trying to delete {}, and {} with size of {}",
+                    metaFileToRemove::getName,
+                    contentFileToRemove::getName,
+                    () -> metaFileToRemove.length() + contentFileToRemove.length());
+                this.openFiles.asMap().remove(metaFileToRemove.getName());
+                this.openFiles.asMap().remove(contentFileToRemove.getName());
+//                this.openFiles.invalidate(metaFileToRemove.getName());
+//                this.openFiles.invalidate(contentFileToRemove.getName());
+                this.openFiles.cleanUp();
+                return
+                    (!metaFileToRemove.exists() || metaFileToRemove.delete()) &&
+                        (!contentFileToRemove.exists() || contentFileToRemove.delete());
+            })
+            .filter(b -> b)
+            .findFirst()
+            .orElse(Boolean.FALSE);
+    }
+
+    void cleanUp() {
+        this.openFiles.cleanUp();
+    }
+
+    public synchronized void close() {
+        this.openFiles.asMap().forEach((name, file) -> closeFile(name, file, RemovalCause.EXPLICIT));
+        this.openFiles.invalidateAll();
+        this.scheduledExecutorService.shutdownNow();
     }
 }
