@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c) 2020-2021 Mediatheken DLNA Bridge Authors.
+ * Copyright (c) 2020-2023 Mediatheken DLNA Bridge Authors.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -26,13 +26,16 @@ package de.corelogics.mediaview.service.proxy.downloader;
 
 import de.corelogics.mediaview.config.MainConfiguration;
 import de.corelogics.mediaview.util.HttpUtils;
+import jakarta.servlet.http.HttpServletResponse;
+import lombok.Getter;
+import lombok.val;
 import okhttp3.ConnectionPool;
 import okhttp3.OkHttpClient;
 import okhttp3.Request;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.jetbrains.annotations.NotNull;
 
-import javax.servlet.http.HttpServletResponse;
 import java.io.Closeable;
 import java.io.EOFException;
 import java.io.IOException;
@@ -44,6 +47,8 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
+
+import static java.util.Optional.ofNullable;
 
 class ClipDownloader implements Closeable {
     private static final double REQUIRED_MB_PER_SECONDS = 1.3D;
@@ -57,31 +62,33 @@ class ClipDownloader implements Closeable {
     private final Logger logger = LogManager.getLogger(ClipDownloader.class);
     private final MainConfiguration mainConfiguration;
     private final CacheDirectory cacheDir;
+
+    @Getter
     private final String url;
     private final String clipId;
     private final int numParallelConnections;
     private final OkHttpClient httpClient;
-    private ClipMetadata metadata;
+    private final ClipMetadata metadata;
     private BitSet chunksAvailableForDownload;
     private int lastReadInChunk = 0;
     private boolean stopped = false;
 
     public ClipDownloader(
-            MainConfiguration mainConfiguration,
-            CacheDirectory cacheDir,
-            String clipId,
-            String url) throws UpstreamNotFoundException, UpstreamReadFailedException, CacheSizeExhaustedException {
+        MainConfiguration mainConfiguration,
+        CacheDirectory cacheDir,
+        String clipId,
+        String url) throws UpstreamNotFoundException, UpstreamReadFailedException, CacheSizeExhaustedException {
         this.mainConfiguration = mainConfiguration;
         this.cacheDir = cacheDir;
         this.url = url;
         this.clipId = clipId;
         this.numParallelConnections = mainConfiguration.cacheParallelDownloadsPerVideo();
         this.httpClient = new OkHttpClient.Builder()
-                .connectionPool(new ConnectionPool(1, 10, TimeUnit.SECONDS))
-                .callTimeout(10, TimeUnit.SECONDS)
-                .readTimeout(10, TimeUnit.SECONDS)
-                .connectTimeout(5, TimeUnit.SECONDS)
-                .build();
+            .connectionPool(new ConnectionPool(1, 10, TimeUnit.SECONDS))
+            .callTimeout(10, TimeUnit.SECONDS)
+            .readTimeout(10, TimeUnit.SECONDS)
+            .connectTimeout(5, TimeUnit.SECONDS)
+            .build();
         logger.debug("Starting download for {}", this.url);
         this.metadata = loadOrFetchMetaData();
         logger.debug("Initialized metadata to {}", this.metadata);
@@ -92,36 +99,36 @@ class ClipDownloader implements Closeable {
     }
 
     public synchronized void onChunkReceived(String connectionId, ClipChunk clipChunk, byte[] bytes, long timeMs) {
-        logger.debug("Chunk {} received: {} bytes in {} ms (by {})", clipChunk.getChunkNumber(), bytes.length, timeMs, connectionId);
+        logger.debug("Chunk {} received: {} bytes in {} ms (by {})", clipChunk.chunkNumber(), bytes.length, timeMs, connectionId);
         if (!stopped) {
             try {
-                cacheDir.writeContent(this.clipId, clipChunk.getChunkNumber() * CHUNK_SIZE_BYTES, bytes);
+                cacheDir.writeContent(this.clipId, clipChunk.chunkNumber() * CHUNK_SIZE_BYTES, bytes);
             } catch (final IOException e) {
                 logger.warn("Could not write to content file.", e);
             }
-            this.metadata.getBitSet().set(clipChunk.getChunkNumber());
+            this.metadata.bitSet().set(clipChunk.chunkNumber());
             this.updateMetadataFile();
         }
     }
 
     public synchronized void onChunkError(String connectionId, ClipChunk clipChunk, Exception e) {
-        logger.debug("Chunk {} not received due to {} (by {})", clipChunk.getChunkNumber(), e.getMessage(), connectionId);
-        this.chunksAvailableForDownload.set(clipChunk.getChunkNumber());
+        logger.debug("Chunk {} not received due to {} (by {})", clipChunk.chunkNumber(), e.getMessage(), connectionId);
+        this.chunksAvailableForDownload.set(clipChunk.chunkNumber());
     }
 
     public synchronized Optional<ClipChunk> nextChunk(String connectionId) {
         if (!stopped) {
             var chunkNum = this.chunksAvailableForDownload.nextSetBit(lastReadInChunk);
-            if (this.chunksAvailableForDownload.get(this.metadata.getNumberOfChunks() - 1)) {
-                // first DL latest chunk, for meta data queried by various players
-                chunkNum = this.metadata.getNumberOfChunks() - 1;
+            if (this.chunksAvailableForDownload.get(this.metadata.numberOfChunks() - 1)) {
+                // first DL latest chunk, for metadata queried by various players
+                chunkNum = this.metadata.numberOfChunks() - 1;
             }
-            if (chunkNum < 0 || chunkNum >= this.metadata.getNumberOfChunks()) {
+            if (chunkNum < 0 || chunkNum >= this.metadata.numberOfChunks()) {
                 chunkNum = this.chunksAvailableForDownload.nextSetBit(0);
             }
-            if (chunkNum >= 0 && chunkNum < this.metadata.getNumberOfChunks()) {
+            if (chunkNum >= 0 && chunkNum < this.metadata.numberOfChunks()) {
                 this.chunksAvailableForDownload.clear(chunkNum);
-                var chunk = new ClipChunk(chunkNum, chunkNum * CHUNK_SIZE_BYTES, (1 + chunkNum) * CHUNK_SIZE_BYTES - 1);
+                val chunk = new ClipChunk(chunkNum, chunkNum * CHUNK_SIZE_BYTES, (1 + chunkNum) * CHUNK_SIZE_BYTES - 1);
                 logger.debug("handing out {} to {}", chunk, connectionId);
                 return Optional.of(chunk);
             }
@@ -132,42 +139,38 @@ class ClipDownloader implements Closeable {
 
     public synchronized void onConnectionTerminated(String connectionId) {
         logger.debug("connection {} terminated", connectionId);
-        this.connections.remove(connectionId);
-    }
-
-    public String getUrl() {
-        return this.url;
+        ofNullable(this.connections.remove(connectionId)).ifPresent(ClipDownloadConnection::close);
     }
 
     private synchronized void ensureDownloadersPresent() {
-        while (this.connections.size() < numParallelConnections && this.chunksAvailableForDownload.nextSetBit(0) < this.metadata.getNumberOfChunks()) {
-            var connectionId = "dl-thrd-" + currentConnectionId.incrementAndGet();
+        while (this.connections.size() < numParallelConnections && this.chunksAvailableForDownload.nextSetBit(0) < this.metadata.numberOfChunks()) {
+            val connectionId = "dl-thrd-" + currentConnectionId.incrementAndGet();
             logger.debug("Starting new connection {}", connectionId);
             this.connections.put(connectionId, new ClipDownloadConnection(
-                    this, mainConfiguration,
-                    connectionId,
-                    CHUNK_SIZE_BYTES,
-                    REQUIRED_MB_PER_SECONDS / numParallelConnections));
+                this, mainConfiguration,
+                connectionId,
+                CHUNK_SIZE_BYTES,
+                REQUIRED_MB_PER_SECONDS / numParallelConnections));
             this.connections.get(connectionId).start();
         }
     }
 
     private void initializeBitsets() {
-        if (null == this.metadata.getBitSet()) {
-            var bitsetSize = 1 + (int) (this.metadata.getSize() / CHUNK_SIZE_BYTES);
+        if (null == this.metadata.bitSet()) {
+            val bitsetSize = 1 + (int) (this.metadata.size() / CHUNK_SIZE_BYTES);
             logger.debug("creating new bitset of size {}", bitsetSize);
-            this.metadata.setNumberOfChunks(bitsetSize);
-            this.metadata.setBitSet(new BitSet(this.metadata.getNumberOfChunks()));
+            this.metadata.numberOfChunks(bitsetSize);
+            this.metadata.bitSet(new BitSet(this.metadata.numberOfChunks()));
         }
-        this.chunksAvailableForDownload = new BitSet(this.metadata.getNumberOfChunks());
-        this.chunksAvailableForDownload.set(0, this.metadata.getNumberOfChunks());
-        this.chunksAvailableForDownload.xor(this.metadata.getBitSet());
+        this.chunksAvailableForDownload = new BitSet(this.metadata.numberOfChunks());
+        this.chunksAvailableForDownload.set(0, this.metadata.numberOfChunks());
+        this.chunksAvailableForDownload.xor(this.metadata.bitSet());
     }
 
     private void growContentFile() throws CacheSizeExhaustedException {
-        logger.debug("setting content file to length {}", this.metadata::getSize);
+        logger.debug("setting content file to length {}", this.metadata::size);
         try {
-            this.cacheDir.growContentFile(this.clipId, this.metadata.getSize());
+            this.cacheDir.growContentFile(this.clipId, this.metadata.size());
         } catch (final IOException e) {
             throw new RuntimeException(e);
         }
@@ -186,7 +189,7 @@ class ClipDownloader implements Closeable {
 
     private ClipMetadata loadOrFetchMetaData() throws UpstreamNotFoundException, UpstreamReadFailedException {
         try {
-            var fromFile = this.cacheDir.loadMetadata(clipId);
+            val fromFile = this.cacheDir.loadMetadata(clipId);
             if (fromFile.isPresent()) {
                 return fromFile.get();
             }
@@ -199,18 +202,18 @@ class ClipDownloader implements Closeable {
     private ClipMetadata fetchMetadataFromUrl() throws UpstreamNotFoundException, UpstreamReadFailedException {
         logger.debug("Loading metadata via HEAD request from {}", this.url);
         try {
-            var request = HttpUtils.enhanceRequest(
+            val request = HttpUtils.enhanceRequest(
                     this.mainConfiguration,
                     new Request.Builder()
-                            .url(this.url)
-                            .head())
-                    .build();
-            try (var response = this.httpClient.newCall(request).execute()) {
+                        .url(this.url)
+                        .head())
+                .build();
+            try (val response = this.httpClient.newCall(request).execute()) {
                 if (response.isSuccessful()) {
-                    logger.debug("successfull HEAD request with headers: {} for HEAD {}", response.headers(), url);
-                    var meta = new ClipMetadata();
-                    meta.setContentType(response.header("Content-Type"));
-                    meta.setSize(Long.parseLong(response.header("Content-Length")));
+                    logger.debug("successful HEAD request with headers: {} for HEAD {}", response.headers(), url);
+                    val meta = new ClipMetadata();
+                    meta.contentType(response.header("Content-Type"));
+                    meta.size(Long.parseLong(response.header("Content-Length")));
                     return meta;
                 }
                 if (response.code() == HttpServletResponse.SC_NOT_FOUND) {
@@ -231,19 +234,19 @@ class ClipDownloader implements Closeable {
     }
 
     public InputStream openInputStreamStartingFrom(long position, Duration readTimeout) throws EOFException {
-        if (position < 0 || position > metadata.getSize()) {
-            throw new EOFException(String.format("Position %d outside of allowed range: [0-%d]", position, metadata.getSize()));
+        if (position < 0 || position > metadata.size()) {
+            throw new EOFException(String.format("Position %d outside of allowed range: [0-%d]", position, metadata.size()));
         }
         return new InputStream() {
             long currentPosition = position;
 
             @Override
             public int read() throws IOException {
-                var timeoutAt = System.currentTimeMillis() + readTimeout.toMillis();
+                val timeoutAt = System.currentTimeMillis() + readTimeout.toMillis();
                 while (System.currentTimeMillis() < timeoutAt) {
-                    var chunkNo = (int) (currentPosition / CHUNK_SIZE_BYTES);
+                    val chunkNo = (int) (currentPosition / CHUNK_SIZE_BYTES);
                     updateLastReadChunk(chunkNo);
-                    if (metadata.getBitSet().get(chunkNo)) {
+                    if (metadata.bitSet().get(chunkNo)) {
                         return cacheDir.readContentByte(clipId, currentPosition++);
                     } else {
                         try {
@@ -257,20 +260,20 @@ class ClipDownloader implements Closeable {
             }
 
             @Override
-            public int read(byte[] b, int off, int len) throws IOException {
+            public int read(@NotNull byte[] b, int off, int len) throws IOException {
                 // we only read to the chunk limit...
-                var timeoutAt = System.currentTimeMillis() + readTimeout.toMillis();
+                val timeoutAt = System.currentTimeMillis() + readTimeout.toMillis();
                 while (System.currentTimeMillis() < timeoutAt) {
-                    var chunkNo = (int) (currentPosition / CHUNK_SIZE_BYTES);
-                    if (chunkNo >= metadata.getNumberOfChunks()) {
-                        logger.debug("Read position {}} is beyond size of {}}", currentPosition, metadata.getSize());
+                    val chunkNo = (int) (currentPosition / CHUNK_SIZE_BYTES);
+                    if (chunkNo >= metadata.numberOfChunks()) {
+                        logger.debug("Read position {}} is beyond size of {}}", currentPosition, metadata.size());
                         return -1;
                     }
                     updateLastReadChunk(chunkNo);
-                    if (metadata.getBitSet().get(chunkNo)) {
-                        var availableBytesInChunk = (chunkNo + 1) * CHUNK_SIZE_BYTES - currentPosition;
-                        var toRead = (int) Math.min(len, availableBytesInChunk);
-                        var readBytesFromFile = cacheDir.readContentBytes(clipId, currentPosition, b, off, toRead);
+                    if (metadata.bitSet().get(chunkNo)) {
+                        val availableBytesInChunk = (chunkNo + 1) * CHUNK_SIZE_BYTES - currentPosition;
+                        val toRead = (int) Math.min(len, availableBytesInChunk);
+                        val readBytesFromFile = cacheDir.readContentBytes(clipId, currentPosition, b, off, toRead);
                         currentPosition += readBytesFromFile;
                         return readBytesFromFile;
                     } else {
