@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c) 2020-2021 Mediatheken DLNA Bridge Authors.
+ * Copyright (c) 2020-2023 Mediatheken DLNA Bridge Authors.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -25,6 +25,7 @@
 package de.corelogics.mediaview.util;
 
 import io.whitfin.siphash.SipHasher;
+import lombok.val;
 
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
@@ -35,19 +36,18 @@ import java.util.Base64;
 public class HashingUtils {
     // test vector from siphash24.c at https://www.131002.net/siphash/siphash24.c
     private static final byte[] SIP42_KEY_BYTES = ByteBuffer
-            .allocate(16)
-            .order(ByteOrder.LITTLE_ENDIAN)
-            .putLong(0x0706050403020100L)
-            .putLong(0x0f0e0d0c0b0a0908L)
-            .array();
+        .allocate(16)
+        .order(ByteOrder.LITTLE_ENDIAN)
+        .putLong(0x0706050403020100L)
+        .putLong(0x0f0e0d0c0b0a0908L)
+        .array();
 
     public static String idHash(String... idStrings) {
-        var hasher = SipHasher.init(SIP42_KEY_BYTES, 2, 4);
+        val hasher = SipHasher.init(SIP42_KEY_BYTES, 2, 4);
         Arrays.stream(idStrings)
-                .map(s -> s.getBytes(StandardCharsets.UTF_8))
-                .forEach(hasher::update);
-
-        ByteBuffer b = ByteBuffer.allocate(8).order(ByteOrder.LITTLE_ENDIAN);
+            .map(s -> s.getBytes(StandardCharsets.UTF_8))
+            .forEach(hasher::update);
+        final var b = ByteBuffer.allocate(8).order(ByteOrder.LITTLE_ENDIAN);
         b.putLong(hasher.digest());
         b.rewind();
         return Base64.getEncoder().withoutPadding().encodeToString(b.array());
